The Great Code Cleanup of 2016

+++ Calendula/src/main/java/es/usc/citius/servando/calendula/activities/ReminderNotification.java
@@ -13,7 +13,7 @@
  *    GNU General Public License for more details.
  *
  *    You should have received a copy of the GNU General Public License
- *    along with this software.  If not, see <http://www.gnu.org/licenses/>.
+ *    along with this software.  If not, see <http://www.gnu.org/licenses>.
  */
 
 package es.usc.citius.servando.calendula.activities;
@@ -319,7 +319,7 @@ public class ReminderNotification {
 
         if (delayMinutes > 0 && !lost)
         {
-            String repeatTime = DateTime.now().plusMinutes((int) delayMinutes).toString("kk:mm");
+            String repeatTime = DateTime.now().plusMinutes(delayMinutes).toString("kk:mm");
             style.setSummaryText(context.getString(R.string.notification_repeat_message, repeatTime));
         } else
         {
